be:added user

+++ Backend/QuizApp/Repositories/IBaseRepository.cs
@@ -8,11 +8,11 @@ namespace QuizApp.Repositories
         Task<List<T>> GetAll(params Expression<Func<T, object>>[] includes);
         //Helps with the query some data where relationship comes in??
 
-        Task<T> Get(int id, params Expression<Func<T, object>>[] includes);
+        Task<T> Get(long id, params Expression<Func<T, object>>[] includes);
 
 
         Task<T> Add(T entity); //where this generates, help to create 
         Task<T> Update(T entity);
-        Task<T> Delete(int id);
+        Task<T> Delete(long id);
     }
 }
